Add a robots.txt file
