+++ data/configRobots.js
@@ -1,0 +1,14 @@
+const config = {
+  siteTitle: "Impulse",
+  siteDescription:
+    "Conceitos de design com inovação digital para criar plataformas web e mobile fáceis de usar e que trazem resultados.",
+  siteTitleShort: "IP",
+  siteRss: "/rss.xml",
+  siteImage: "",
+  googleAnalyticsID: "UA-141117092-12",
+  themeColor: "#000000",
+  backgroundColor: "#fff",
+  siteUrl: "http://impulse.works",
+}
+
+module.exports = config
